Put setting types in its own file
